modified:   app/Http/Controllers/api/ShearController.php 	modified:   database/migrations/2025_01_31_185221_create_shears_table.php 	modified:   database/migrations/2025_01_31_185517_create_matrices_table.php 	modified:   resources/js/components/Shear.vue 	modified:   routes/web.php

+++ database/migrations/2025_01_31_185517_create_matrices_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->timestamps();
             $table->unsignedBigInteger('edge_id')->nullable();
-            $table->unsignedBigInteger('c')->nullable();
+            $table->float('c')->nullable();
         });
     }
